requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
biopython>=1.83
matplotlib>=3.7.0
transformers>=4.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
PyYAML>=6.0.0
accelerate>=0.26.0
